pyproject.toml
[project]
name = "assessment-agent"
version = "0.1.0"
description = "AI agent that conducts conversational assessments of learner understanding in fractions"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.79.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
assess = "src.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
